chore: revert results page logic
